requirements.txt
# base {{ project_name|title }} requirements
Django==1.4.3
Fabric==1.5.2
South==0.7.6
Unipath==0.2.1
distribute==0.6.27
django-appconf==0.5
django-compressor==1.2
-e git://github.com/jbalogh/django-debug-cache-panel@2ff3c45be962e08f02928d021c61e7008964020a#egg=django_debug_cache_panel-dev
django-debug-toolbar==0.9.4
django-extensions==1.0.2
gunicorn==0.17.2
paramiko==1.9.0
psycopg2==2.4.6
pycrypto==2.6
wsgiref==0.1.2
